Json update and text mishmash improvement
Applying the parsing method of text from a Json webRequest and the Post method into the new script InfoDisplayExtra instead of non. Changed and improved the way how the json file parses information and posts it inside the text object. Also did text alinement(some basic stuff) beside coding the scripts and editing.

+++ PROJECT NOT DELETE TEST/Assets/Scripts/Serializable Classes/Hotspot.cs
@@ -1,9 +1,10 @@
 [System.Serializable]
-public class Hotspot {
+public class Hotspot
+{
+    // public string _id;
     public string name;
-    public byte scaleX;
-    public byte scaleY;
-    public byte scaleZ;
+    public float scaleX;
+    public float scaleY;
+    public float scaleZ;
     public Image[] images;
-
 }
